src/app/home/home.component.spec.ts
import { HttpClient } from '@angular/common/http';
import {
  HttpClientTestingModule,
  HttpTestingController,
} from '@angular/common/http/testing';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { WeatherService } from '../services/weather.service';
import { HomeComponent } from './home.component';

describe('HomeComponent', () => {
  let component: HomeComponent;
  let fixture: ComponentFixture<HomeComponent>;
  //let weatherService: WeatherService;
  let httpClient: HttpClient;
  let httpMock: HttpTestingController;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [HomeComponent],
      providers: [WeatherService],
      imports: [HttpClientTestingModule],
    }).compileComponents();
    //weatherService = TestBed.inject(WeatherService);
    //httpClient = TestBed.inject(HttpClient);
    httpMock = TestBed.get(HttpTestingController);
  });

  beforeEach(() => {
    //weatherService = TestBed.get(WeatherService);
    fixture = TestBed.createComponent(HomeComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  /*
  it('should change message to hello if clicked', async() => {
    const fixture = TestBed.createComponent(HomeComponent);
    fixture.detectChanges();
    expect(fixture.componentInstance.message).toBe("");
    fixture.debugElement.nativeElement.querySelector('button').click();
    expect(fixture.componentInstance.message).toBe("Hello");
  });

  it('should do something', async() => {
    const fixture = TestBed.createComponent(HomeComponent);
    fixture.detectChanges();
    expect(fixture.componentInstance.weather).toBeGreaterThan(1);
  });
  */
});
